requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
dnspython==2.2.1
frozenlist==1.3.0
idna==3.3
imagekitio==2.2.8
motor==2.5.1
multidict==6.0.2
numpy==1.22.3
opencv-python==4.5.5.64
passlib==1.7.4
Pillow==9.1.0
pymongo==3.12.3
pyTelegramBotAPI==4.4.0
python-decouple==3.6
pytz==2022.1
requests==2.27.1
soupsieve==2.3.1
urllib3==1.26.9
yarl==1.7.2
